# public/dist/swagger.yaml
swagger: '2.0'
info:
  description: API Documentation
  version: 1.0.0
  title: FoodPhone.menu
host: 'localhost:8080'
basePath: /api
...
schemes:
  - http
paths:
...
  /category:
    post:
      tags:
        - category
      summary: Create new category
      description: ""
      operationId: createCategory
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: new category
          required: false
          schema:
            $ref: "#/definitions/Category"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/Category"
        "400":
          description: Invalid Category
  /category/{categoryId}:
    get:
      tags:
        - category
      summary: Get specific category
      description: ""
      operationId: getCategoryById
      produces:
        - application/json
      parameters:
        - name: categoryId
          in: path
          description: ID of category
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Category'
        '400':
          description: Invalid ID supplied
        '404':
          description: Category not found
    put:
      tags:
        - category
      summary: Update an existing category
      description: ""
      operationId: updateCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: categoryId
          in: path
          description: ID of category
          required: true
          type: string
        - in: body
          name: body
          description: Update category
          required: false
          schema:
            $ref: "#/definitions/Category"
      responses:
        "405":
          description: Validation exception
        "404":
          description: Category not found
        "400":
          description: Invalid ID supplied
    delete:
      tags:
        - category
      summary: Delete category by ID
      description: ""
      operationId: deleteCategory
      produces:
        - application/json
      parameters:
        - name: categoryId
          in: path
          description: ID of the category to be deleted
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Category'
        '400':
          description: Invalid ID supplied
        '404':
          description: Category not found

#rol
...
  /menuCategory:
    post:
      tags:
        - menuCategory
      summary: Create new menuCategory
      description: ""
      operationId: createMenuCategory
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: new menuCategory
          required: false
          schema:
            $ref: "#/definitions/MenuCategory"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/MenuCategory"
        "400":
          description: Invalid MenuCategory
  /menuCategory/{menuCategoryId}:
    get:
      tags:
        - menuCategory
      summary: Get specific menuCategory
      description: ""
      operationId: getMenuCategoryById
      produces:
        - application/json
      parameters:
        - name: menuCategoryId
          in: path
          description: ID of menuCategory
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/MenuCategory'
        '400':
          description: Invalid ID supplied
        '404':
          description: MenuCategory not found
    put:
      tags:
        - menuCategory
      summary: Update an existing menuCategory
      description: ""
      operationId: updateMenuCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: menuCategoryId
          in: path
          description: ID of menuCategory
          required: true
          type: string
        - in: body
          name: body
          description: Update menuCategory
          required: false
          schema:
            $ref: "#/definitions/MenuCategory"
      responses:
        "405":
          description: Validation exception
        "404":
          description: MenuCategory not found
        "400":
          description: Invalid ID supplied
    delete:
      tags:
        - menuCategory
      summary: Delete menuCategory by ID
      description: ""
      operationId: deleteMenuCategory
      produces:
        - application/json
      parameters:
        - name: menuCategoryId
          in: path
          description: ID of the menuCategory to be deleted
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/MenuCategory'
        '400':
          description: Invalid ID supplied
        '404':
          description: MenuCategory not found
          
#company
...
  /company:
    post:
      tags:
        - company
      summary: Create new company
      description: ""
      operationId: createCompany
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: new company
          required: false
          schema:
            $ref: "#/definitions/Company"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/Company"
        "400":
          description: Invalid Company
  /company/{companyId}:
    get:
      tags:
        - company
      summary: Get specific company
      description: ""
      operationId: getCompanyById
      produces:
        - application/json
      parameters:
        - name: companyId
          in: path
          description: ID of company
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Company'
        '400':
          description: Invalid ID supplied
        '404':
          description: Company not found
    put:
      tags:
        - company
      summary: Update an existing company
      description: ""
      operationId: updateCompany
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: companyId
          in: path
          description: ID of company
          required: true
          type: string
        - in: body
          name: body
          description: Update company
          required: false
          schema:
            $ref: "#/definitions/Company"
      responses:
        "405":
          description: Validation exception
        "404":
          description: Company not found
        "400":
          description: Invalid ID supplied
    delete:
      tags:
        - company
      summary: Delete company by ID
      description: ""
      operationId: deleteCompany
      produces:
        - application/json
      parameters:
        - name: companyId
          in: path
          description: ID of the company to be deleted
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Company'
        '400':
          description: Invalid ID supplied
        '404':
          description: Company not found

#user
...
  /parameter:
    post:
      tags:
        - parameter
      summary: Create new Parameter
      description: ""
      operationId: createParameter
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New Parameter
          required: false
          schema:
            $ref: "#/definitions/Parameter"
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/Parameter"
        "400":
          description: Invalid Parameter
  /parameter/{parameterId}:
    get:
      tags:
        - parameter
      summary: Get specific Parameter
      description: ""
      operationId: getParameterById
      produces:
        - application/json
      parameters:
        - name: parameterId
          in: path
          description: ID of Parameter
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Parameter'
        '400':
          description: Invalid ID supplied
        '404':
          description: Parameter not found
    put:
      tags:
        - parameter
      summary: Update an existing Parameter
      description: ""
      operationId: updateParameter
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: parameterId
          in: path
          description: ID of parameter
          required: true
          type: string
        - in: body
          name: body
          description: Update Parameter
          required: false
          schema:
            $ref: "#/definitions/Parameter"
      responses:
        "405":
          description: Validation exception
        "404":
          description: Parameter not found
        "400":
          description: Invalid ID supplied
    delete:
      tags:
        - parameter
      summary: Delete Parameter by ID
      description: ""
      operationId: deleteParameter
      produces:
        - application/json
      parameters:
        - name: parameterId
          in: path
          description: ID of the Parameter to be deleted
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Parameter'
        '400':
          description: Invalid ID supplied
        '404':
          description: Parameter not found
definitions:
...
  Category:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      isActive:
        type: boolean

#rol
...
  MenuCategory:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      isActive:
        type: boolean

#company
  Company:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      details:
        type: string
      categories:
        type: array
        items:
          type: string
      isDelivery:
        type: boolean
      isCash:
        type: boolean
      isCreditCard:
        type: boolean
      isActive:
        type: boolean
      isOpenNow:
        type: boolean
      timetable:
        type: array
        items:
          $ref: "#/definitions/CompanyTimeTable"
  CompanyTimeTable:
    type: object
    properties:
      day:
        type: string
      open:
        type: string
      close:
        type: string
      isClose:
        type: boolean

#user
...
  Parameter:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      value:
        type: string
      isActive:
        type: boolean
      details:
        type: array
        items:
          $ref: "#/definitions/ParameterDetails"
  ParameterDetails:
    type: object
    properties:
      name:
        type: string
      details:
        type: string
      numberId:
        type: number
      value:        
        type: string
      order:
        type: number
      isActive:
        type: boolean

#questionType
...